refactor([issue-742]): split agentLifecycle.js monolith into focused modules
agentLifecycle.js had grown to 1744 LOC. Extract the internal blocks of its
two giant orchestrators into self-contained modules, leaving the public
orchestrators (spawnAgentForTask / handleAgentCompletion) as thin coordinators:

- agentProviderResolution.js — provider availability/fallback + user override
  + per-task model selection (resolveAgentProviderAndModel)
- agentWorkspacePrep.js — workspace path + git pull + JIRA branch + worktree
  provisioning, returning a discriminated ready/deferred/blocked outcome
- agentWorktreeCleanup.js — cleanupAgentWorktree + spawnReviewLoopFollowUp +
  spawnMergeRecoveryTask (leaf cluster, breaks the would-be import cycle)
- agentCompletionCleanup.js — handlePipelineProgression + the post-finalize
  cleanup orchestration (runAgentCompletionCleanup)

agentLifecycle.js drops to 856 LOC and re-exports the moved functions so
existing consumers (subAgentSpawner, agentManagement) keep importing them from
there. Adds unit tests for the two new branching modules.
